Remove copy from hon
